A [DBI]-style database connection to GBIF data

Returns a database connection to the local GBIF parquet file.

gbif_conn(
  dir = gbif_parquet_dir(version = gbif_version(local = TRUE)),
  tblname = "gbif",
  backend = c("arrow", "duckdb")
)

Arguments

dir

the directory containing all parquet files to be read

tblname

name of the table to be created in the duckdb VIEW

backend

Use arrow or duckdb as backend connection?

Value

a DBIconnection object

Examples

if (FALSE) { # interactive()

gbif.parquet <- gbif_example_data()
con <- gbif_conn(gbif.parquet)
}